Fixing out of order messages with a priority queue
The client uses a priorityQueue from an old version of async.

For some reason this implementation puts items with the same
priority ahead of existing tasks with the same priority.

See: https://github.com/caolan/async/blob/621f13805aa326865b85dbbf7128baf7146ab976/lib/async.js#L1030

Since this is not being used as a priority queue anyways(no priority
is being passed to the queue) we can fix this by simply switching
to a normal queue. Problem solved :)

Adding a test for the problem as well. Ensures that code is executed
in order.