WIP: webclient types clean up
the goal of this commit is to remove the @ts-ignore found in WebClient#apiCall(). it
turns out that accomplishing this is best done by making upstream changes to the
`@types/got` package. the comment is revised to leave a trail to the relevent info.

while working towards this goal it became evident that the FormCanBeURLEncoded is a
confusing name, and URLEncodableForm is a better name, so its renamed to that. also,
the BodyCanBeFormMultipart interface has a confusing name, and doesn't provide any
value over Readable, since got's options are typed in terms of Readable, so its been
removed. if in the future there's a need to further specify the type that represents
a form multipart body, we can add that in the future.