uses travis test matrix and codecov flags to get better CI reporting
the strategy is to use a combination of environment variables with
lerna's scope argument to get individual builds for each package.
this depends on each package naming its scripts precisely the same
thing for certain operations: test, lint, and coverage. notably,
its okay for a script to be missing because lerna run will handle
only dispatching to packages where the script is defined.