test(WebClient): add test cases for invalid Retry-After headers
This adds test cases for when there's no Retry-After header on a 429
response and for when there's an invalid Retry-After header on a 429
response.

This also includes a test case to assert that buildResult includes
retrySec in the response_metadata object of the data response. This
test case is kind of finicky though, because it looks like in the
case of a 429 response, that particular block isn't run because
the makeRequest function either throws an error if rejectRateLimitedCalls
is true or waits and trys again, in which case there won't be a Retry-After
header on the non-429 response after the retry timeout ends.