fix(package.json): update form-data to 2.5.0, removes @types/form-data
I noticed that the Travis CI build was failing due to a missing type
definition for form-data in the WebAPI package. It looks like the
current release of form-data includes its own type-definitions, which
means that @types/form-data isn't necessary anymore.

Removing @types/form-data and updating form-data to 2.5.0 fixed the
lerna bootstrap command.