fix: uses separate jwt for device sync
My intention was to only send the random byte array as the value
for the OAuth state param, and for the JWT to only exist in the
cookie. However, that represents a potentially complex and/or
breaking change. However, if the same JWT is used as the state
param and in the cookie, and adversary who is able to capture the
redirect url can easily spoof a synchronized device. So the examples
sign JWTs specifically for the device, and provide the synchronizer
to the callbackHandler for comparison.