oauth: correct nodejs `.url` usage
this is an [IncomingMessage](https://nodejs.org/docs/latest-v12.x/api/http.html#http_class_http_incomingmessage) object which gives the full URL for `.url`. We either want to switch to getting path, or check that we match the start of the URL. Only putting what worked for me in dev/testing.